Read character by character
I did a bunch of wizardry to turn off canonical mode and echoing. Then implemented an array of characters typed so that it can be treated as typing line-by-line. I also added a list of commands run in the REPL that the up and down arrow keys can go through.

What's not working: delete/backspace, left/right arrow keys. These have gotten worse since before this change, and need to be fixed.

Closes #6